requirements.txt
apache-airflow-providers-sftp
apache-airflow-providers-ssh